chore: Use summary level log when detail level flags is invalid
When an invalid detail level is provided, this function returns a
default value (resource.DetailLevelDetailed) along with an error. This
can be misleading, as a caller might ignore the error and use the
returned default value. It's better practice to return the type's zero
value when an error occurs. The zero value for resource.DetailLevel is
resource.DetailLevelSummary.